+++ app/src/main/java/com/techcourse/session/UserSession.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller;
+package com.techcourse.session;
 
 import com.techcourse.domain.User;
 import jakarta.servlet.http.HttpSession;
