+++ mvc/src/main/java/nextstep/mvc/servlet/DispatcherServlet.java
@@ -1,4 +1,4 @@
-package nextstep.mvc;
+package nextstep.mvc.servlet;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
